tehtävä 4.21

+++ osa4/blogilista/controllers/blogs.js
@@ -3,20 +3,20 @@ const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
 const User = require("../models/user");
 
+
 blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({}).populate("user", {username: 1, name: 1});
+  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
   response.json(blogs);
 });
-
-
 
 blogsRouter.post("/", async (request, response) => {
   const body = request.body;
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
 
   if (!decodedToken.id) {
-    return response.status(401).json({error: "token missing or invalid"})
+    return response.status(401).json({ error: "token missing or invalid" })
   }
+
   const user = await User.findById(decodedToken.id)
 
   if (body.title === undefined) {
@@ -47,9 +47,22 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
 
-  response.status(204).end();
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  const user = await User.findById(decodedToken.id)
+
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end();
+  } else {
+    response.status(401).json({ error: "tokens not matching" })
+  }
+
 });
 
 blogsRouter.put("/:id", async (request, response) => {
